requirements.txt
pip
setuptools
wheel

python-dotenv
requests

fastapi

pydantic

uvicorn

sqlalchemy

pyjwt

bcrypt

httpx

pytest